axis info for data is integer, changed to float
Normally when dealing with xrdml files, the equipment does not have perfect integer start and end positions in the file. This is the case with all our files. Therefore the array should go from and to floats, like the intervals also are, to represent the proper start and end positions for Theta2.